rename: objects 폴더로 diary 이동 - DiaryModel에서 Diary 제거 - DiaryModel에 임시 dummy data 추가 - DiaryModel에 선택된 Diary 데이터를 저장하는 diary 변수 생성 - DiaryModel을 초기화하는 rest() 함수 생성
